requirements.txt
# ============================================================
# E-commerce Sales Pipeline — Python Dependencies
# ============================================================

# ------ Airflow (pinned; extras needed for PostgreSQL) ------
apache-airflow[postgres]==2.9.1

# ------ Data processing ------
pandas==2.2.2
numpy==1.26.4
pyarrow==16.0.0          # fast parquet I/O for XCom staging
pandera==0.19.3           # DataFrame schema validation

# ------ Database ------
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# ------ Data generation ------
Faker==24.9.0

# ------ Testing ------
pytest==8.2.0
pytest-cov==5.0.0
pytest-mock==3.14.0

# ------ Utilities ------
python-dotenv==1.0.1

# ------ Dashboard (Streamlit showcase) ------
streamlit>=1.35.0
plotly>=5.20.0
